Create a mapper module
We should separate the mapper codebase from the facade module to allow
for more modularity.

The mapper code does not depend on the providers and therefore can be
extracted out.

This allows us to more easily use the mapper code in other modules,
while avoiding cyclic dependencies between modules
