+++ facade/src/main/java/org/jboss/pnc/facade/impl/BrewPusherImpl.java
@@ -25,9 +25,8 @@ import org.jboss.pnc.common.json.ConfigurationParseException;
 import org.jboss.pnc.common.util.StringUtils;
 import org.jboss.pnc.dto.BuildPushResult;
 import org.jboss.pnc.dto.requests.BuildPushRequest;
-import org.jboss.pnc.dto.response.Page;
 import org.jboss.pnc.facade.BrewPusher;
-import org.jboss.pnc.facade.mapper.api.BuildPushResultMapper;
+import org.jboss.pnc.mapper.api.BuildPushResultMapper;
 import org.jboss.pnc.managers.BuildResultPushManager;
 import org.jboss.pnc.managers.Result;
 import org.jboss.pnc.model.BuildRecord;
