+++ facade/src/main/java/org/jboss/pnc/facade/providers/BuildProviderImpl.java
@@ -28,8 +28,8 @@ import org.jboss.pnc.dto.response.Graph;
 import org.jboss.pnc.dto.response.Page;
 import org.jboss.pnc.dto.response.SSHCredentials;
 import org.jboss.pnc.enums.BuildStatus;
-import org.jboss.pnc.facade.mapper.api.BuildConfigurationRevisionMapper;
-import org.jboss.pnc.facade.mapper.api.BuildMapper;
+import org.jboss.pnc.mapper.api.BuildConfigurationRevisionMapper;
+import org.jboss.pnc.mapper.api.BuildMapper;
 import org.jboss.pnc.facade.providers.api.BuildPageInfo;
 import org.jboss.pnc.facade.providers.api.BuildProvider;
 import org.jboss.pnc.facade.validation.EmptyEntityException;
