+++ facade/src/main/java/org/jboss/pnc/facade/providers/ProductMilestoneProviderImpl.java
@@ -21,7 +21,7 @@ import org.jboss.pnc.dto.ProductMilestone;
 import org.jboss.pnc.dto.ProductMilestoneRef;
 import org.jboss.pnc.dto.response.Page;
 import org.jboss.pnc.dto.validation.groups.WhenUpdating;
-import org.jboss.pnc.facade.mapper.api.ProductMilestoneMapper;
+import org.jboss.pnc.mapper.api.ProductMilestoneMapper;
 import org.jboss.pnc.facade.providers.api.ProductMilestoneProvider;
 import org.jboss.pnc.facade.validation.ConflictedEntryException;
 import org.jboss.pnc.facade.validation.ConflictedEntryValidator;
