+++ mapper/src/main/java/org/jboss/pnc/mapper/BrewNameWorkaround.java
@@ -15,7 +15,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.jboss.pnc.facade.mapper;
+package org.jboss.pnc.mapper;
 
 import java.util.HashMap;
 import java.util.Map;
