+++ mapper/src/main/java/org/jboss/pnc/mapper/api/BuildPushResultMapper.java
@@ -15,7 +15,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.jboss.pnc.facade.mapper.api;
+package org.jboss.pnc.mapper.api;
 
 import org.jboss.pnc.dto.BuildPushResult;
 import org.jboss.pnc.model.BuildRecordPushResult;
