+++ mapper/src/main/java/org/jboss/pnc/mapper/api/MapperCentralConfig.java
@@ -15,7 +15,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.jboss.pnc.facade.mapper.api;
+package org.jboss.pnc.mapper.api;
 
 import org.mapstruct.MapperConfig;
 import org.mapstruct.ReportingPolicy;
@@ -27,7 +27,7 @@ import org.mapstruct.ReportingPolicy;
 @MapperConfig(
         unmappedSourcePolicy = ReportingPolicy.ERROR,
         unmappedTargetPolicy = ReportingPolicy.ERROR,
-        implementationPackage = "org.jboss.pnc.facade.mapper",
+        implementationPackage = "org.jboss.pnc.mapper",
         componentModel = "cdi"
 )
 public interface MapperCentralConfig {
